requirements.txt
Flask==2.0.1
Flask_Cors==3.0.10
python-dotenv==0.21.0
SQLAlchemy==1.4.22
Werkzeug==2.0.1
